backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6  # For handling file uploads
pillow==10.0.0
numpy==1.24.3
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
flower==2.0.1  # For monitoring Celery tasks 